feat: Refatorar servidor para utilizar rotas e mover lógica de manipulação de usuários para um novo módulo de rotas
