+++ src/server.js
@@ -1,7 +1,7 @@
 import http from "http"
 import { json } from "./middlewares/json.js"
-import { Database } from "./database.js"
-import { randomUUID } from "node:crypto"
+import { routes } from "./routes.js"
+
 // - Criar usuarios
 // - Listar usuarios
 // - Editar usuarios
@@ -23,29 +23,16 @@ import { randomUUID } from "node:crypto"
 // - Cebeçalhos (Requisição/Resposta) => Metadados
 // - HTTP Status Code
 
-const database = new Database()
-
 const server = http.createServer(async (req, res) => {
   const { method, url } = req
 
   await json(req, res)
 
-  if (method === "GET" && url === "/users") {
-    const users = database.select("users")
-    return res.end(JSON.stringify(users))
-  }
-
-  if (method === "POST" && url === "/users") {
-    const { name, email } = req.body
-
-    const user = {
-      id: randomUUID(),
-      name,
-      email,
-    }
-    database.insert("users", user)
-
-    return res.writeHead(201).end()
+  const route = routes.find((route) => {
+    return route.method === method && route.path === url
+  })
+  if (route) {
+    return route.handler(req, res)
   }
 
   return res.writeHead(404).end()
